pyproject.toml
[tool.poetry]
name = "swap"
version = "0.1.0"
description = "DEX indexer with Apibara"
authors = []

[tool.poetry.scripts]
swap-indexer = "swap.main:cli"

[tool.poetry.dependencies]
python = ">3.8,<3.10"
apibara = "^0.5.16"
click = "^8.1.3"
starknet-py = "^0.6.1a0"
pymongo = {extras = ["srv"], version = "^4.1.1"}
structlog = "^22.3.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.151.0"}
aiohttp = "^3.8.3"
aiohttp_cors = "^0.7.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
